"ads2_m8_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2 assignments/m8/1. Shortest Paths/1. Shortest Paths/Solution.java
@@ -28,7 +28,7 @@ public final class Solution {
 			int source = hashobj.get(srcDest[0]);
 			DijkstraUndirectedSP dijkstra = new DijkstraUndirectedSP(ewgobj, source);
 			if(dijkstra.hasPathTo(hashobj.get(srcDest[1]))) {
-				System.out.println(dijkstra.distTo(hashobj.get(srcDest[1])));
+				System.out.println((int)dijkstra.distTo(hashobj.get(srcDest[1])));
 			}
 			queries--;
 		}
